pyproject.toml
[tool.poetry]
name = "mis-eventos-backend"
version = "0.1.0"
description = "Backend FastAPI para Mis Eventos"
authors = ["Nicolas Esteban Zapata Fontecha"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras=["standard"], version="^0.22.0"}
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.7"
python-jose = "^3.3.0"
passlib = {extras=["bcrypt"], version="^1.7.4"}
python-dotenv = "^1.0.0"
alembic = "^1.12.0"
pydantic = {extras=["email"], version="^1.10.15"}
pytest = "^7.3.1"
httpx = "^0.24.0"
python-multipart = "^0.0.9"
